64bit fixes
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@322 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ include/types.h
@@ -22,7 +22,7 @@ typedef unsigned long	ulong;
 
 #ifndef WIN32
 // WIN32 defines this in windows.h
-typedef int DWORD;
+typedef long DWORD;
 typedef uchar byte;
 #endif
